chore: add shared utils for git secrets tester
